pyproject.toml
[tool.poetry]
name = "pycon2020"
version = "0.1.0"
description = ""
authors = ["Kazuki Watanabe <k.watanabe@jxpress.net>"]

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.68.1"
PyYAML = "^5.4.1"
pytest = "^6.2.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
